+++ application/views/invoice_templates/pdf/InvoiceMuse - paid.php
@@ -1,0 +1,4 @@
+<?php
+
+$invoice_mode = 'paid'; // Need in InvoiceMuse template
+include __DIR__ . '/InvoiceMuse.php';