+++ application/views/quote_templates/public/InvoiceMuse_Web.php
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width,initial-scale=1">
 
     <title>
-        <?php echo get_setting('custom_title', 'InvoicePlane', true); ?>
+        <?php echo get_setting('custom_title', 'InvoiceMuse', true); ?>
         - <?php _trans('quote'); ?> <?php echo htmlsc($quote->quote_number); ?>
     </title>